adapt versioning to start at 0

+++ driver/bash/bash.go
@@ -27,6 +27,6 @@ func (driver *Driver) Migrate(f file.File, pipe chan interface{}) {
 	return
 }
 
-func (driver *Driver) Version() (uint64, error) {
-	return uint64(0), nil
+func (driver *Driver) Version() (int, error) {
+	return 0, nil
 }
